% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wmatrix.R
\docType{data}
\name{Wmatrix}
\alias{Wmatrix}
\title{Proximity matrix}
\format{
A 71*71 proximity matrix of the areas. It must be in row-standerdized form
}
\usage{
data(Wmatrix)
}
\description{
Proximity matrix for the areas included in data set of Chandra et al. (2017)
}
\section{Reference}{

Chandra, H., Salvati, N., & Chambers, R. (2017). Small area prediction of counts under a non-stationary spatial model. Spatial Statistics. 20. 30-56. DOI:10.1016/j.spasta.2017.01.004.
}

\examples{
data(Wmatrix)
}
\keyword{datasets}
